Added Product class with code and added inheritance code
